archive client side complete

+++ src/main/java/by/bylnova/archive/server/service/impl/ArchiveServiceImpl.java
@@ -5,6 +5,7 @@ import by.bylnova.archive.server.dao.DocumentDAO;
 import by.bylnova.archive.server.dao.exception.DAOException;
 import by.bylnova.archive.server.dao.factory.DAOFactory;
 import by.bylnova.archive.server.service.ArchiveService;
+import by.bylnova.archive.server.service.UserSession;
 import by.bylnova.archive.server.service.exception.ServiceException;
 
 import java.util.List;
@@ -12,13 +13,18 @@ import java.util.List;
 public class ArchiveServiceImpl implements ArchiveService {
     private final DAOFactory daoObjectFactory = DAOFactory.getInstance();
     private final DocumentDAO documentDAO = daoObjectFactory.getDocumentDAO();
+    private final UserSession userSession = UserSession.getInstance();
     @Override
     public void addNewDocument(Document document) throws ServiceException {
         if (document == null) {
             throw new ServiceException("Empty document");
         }
         try {
-            documentDAO.addDocument(document);
+            if (userSession.isUserActive()) {
+                if (userSession.getActiveUser().getUserRole().equals("user")) {
+                    documentDAO.addDocument(document);
+                }
+            }
         } catch (DAOException e) {
             throw new ServiceException(e);
         }
@@ -27,7 +33,11 @@ public class ArchiveServiceImpl implements ArchiveService {
     @Override
     public void deleteDocument(long id) throws ServiceException {
         try {
-            documentDAO.deleteDocument(id);
+            if (userSession.isUserActive()) {
+                if (userSession.getActiveUser().getUserRole().equals("admin")) {
+                    documentDAO.deleteDocument(id);
+                }
+            }
         } catch (DAOException e) {
             throw new ServiceException(e);
         }
@@ -41,7 +51,10 @@ public class ArchiveServiceImpl implements ArchiveService {
     @Override
     public List<Document> listAllDocuments() throws ServiceException {
         try {
-            return documentDAO.listDocuments();
+            if (userSession.isUserActive()) {
+                return documentDAO.listDocuments();
+            }
+            return List.of();
         } catch (DAOException e) {
             throw new ServiceException(e);
         }
